Fix : user, memo 수정
이슈 코드 : [S10P12D103]

+++ backend/cherish/src/main/java/com/ssafy/cherish/user/model/dto/UserDto.java
@@ -13,7 +13,7 @@ public class UserDto {
 
     private int id;
     private String idStr;
-    private int kakaoId;
+    private long kakaoId;
     private int coupleId;
     private String nickname;
     private String email;
